Rename ExpressionParser to QueryParser and move to seperate package
